feat: refactor astro-decap-cms integration and add utility functions for preview handling
